Imported Game Jam Assets + Setup
Imported all of the required assets from the game jam unity project and set them up to work. Combined menu main web and win into menu main with the MainMenuManager script deactivating the quit button if the platform is webgl. Also got the input system working again with the player and pause menu. Things should look and function the same as the latest version of the game jam unity project.