Fixed Issue with Volume Settings Resetting on Pause Menu
Fixed the issue with the volume settings getting reset on when the pause menu was first opened. This was because the Get Levels was getting the default values versus from the saved file. To fix this, the game manager now calls initalize on volume manager scritable object when the load next level function is called. This makes sure that the vales from the file are loaded in before get levels is called.